requirements.txt
Flask>=2.3.0,<4.0.0
boto3>=1.28.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
Pillow>=10.0.0
gunicorn>=20.0.0